requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.2
filelock==3.0.12
ibm-cloud-sdk-core==3.10.1
ibm-watson==5.2.2
idna==2.10
jaraco.context==4.0.0
more-itertools==8.7.0
numpy==1.20.2
pipenv==2021.5.29
platformdirs==2.3.0
PyAudio==0.2.11
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.19.0
requests==2.25.1
six==1.16.0
SpeechRecognition==3.8.1
urllib3==1.26.4
virtualenv==20.7.2
virtualenv-clone==0.5.7
websocket-client==1.1.0
wolframalpha==5.0.0
xmltodict==0.12.0